requirements.txt
# Основні залежності
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1

# Машинне навчання
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4

# Обробка тексту
spacy==3.8.2

# База даних
aiosqlite==0.20.0

# Додаткові утиліти
python-multipart==0.0.18

# Для розробки (опціонально)
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1